WIP #1
```yaml
- name: create backuppc user
  user:
    name: "{{ item.username }}"
    shell: "{{ item.shell }}"
    createhome: "yes"
    state: "present"
    home: "{{ item.home }}"
    comment: "{{ item.comment }}"
    password: "*"
  with_items:
      - "{{ backuppc_client }}"
  tags: backuppc
```
https://docs.ansible.com/ansible/latest/modules/user_module.html

```
password
-
Optionally set the user's password to this crypted value.
On macOS systems, this value has to be cleartext. Beware of security issues.
To create a disabled account or Linux systems, set this to '!' or '*'.
See https://docs.ansible.com/ansible/faq.html#how-do-i-generate-crypted-passwords-for-the-user-module for details on various ways to generate these password values.
```